Move initGlobals to storyUtils
