+++ packages/components/src/stories/AssayPropertiesPanel.stories.tsx
@@ -8,7 +8,7 @@ import { Meta, Story } from '@storybook/react/types-6-0';
 
 import { AssayProtocolModel, AssayPropertiesPanel } from '..';
 
-import initGlobal from './initGlobal';
+import { initGlobal } from './storyUtils';
 
 import generalAssayTemplate from '../test/data/assay-getProtocolGeneralTemplate.json';
 import generalAssaySaved from '../test/data/assay-getProtocolGeneral.json';
