+++ packages/components/src/stories/EntityDeleteConfirmModal.stories.tsx
@@ -8,8 +8,7 @@ import { Meta, Story } from '@storybook/react/types-6-0';
 
 import { EntityDeleteConfirmModal, SampleTypeDataType } from '..';
 
-import initGlobal from './initGlobal';
-import { disableControls } from './storyUtils';
+import { disableControls, initGlobal } from './storyUtils';
 
 initGlobal();
