+++ packages/components/src/stories/QueriesListing.stories.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { Meta, Story } from '@storybook/react/types-6-0';
 
 import { QueriesListing } from '..';
-import initGlobal from './initGlobal';
+import { initGlobal } from './storyUtils';
 
 initGlobal();
