+++ packages/components/src/stories/SampleSetSummary.stories.tsx
@@ -17,7 +17,7 @@ import React from 'react';
 import { Meta, Story } from '@storybook/react/types-6-0';
 
 import { SampleSetSummary, User } from '..';
-import initGlobal from './initGlobal';
+import { initGlobal } from './storyUtils';
 
 initGlobal();
