+++ packages/components/src/stories/SchemaListing.stories.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { Meta, Story } from '@storybook/react/types-6-0';
 
 import { SchemaListing } from '..';
-import initGlobal from './initGlobal';
+import { initGlobal } from './storyUtils';
 
 initGlobal();
